fix(mobile): move home to the new tab layout
